add silicon software support
Add libs, and .h's from the silicon software library. This lib is
downloded as an .exe from the Silicon software website for free. We
install it on windows as normal, then I copied the resulting libs and
h's to the ADPCO dir. We then add a Makefile so the .h's get included.
This is the microEnable IV library. You still need to install the
softwae properly because it includes the hardware driver for the card,
that must be registered in the windows OS.  I would include the large
.exe files but free github won't let us.